ci(codeql): isolate deliberate-race lock tests and exclude that one file
The reclaim-identity-guard tests deliberately stage a controlled
open→rm→recreate race at the same lock path to prove the guard survives a
concurrent lock swap. CodeQL's security-extended js/file-system-race (TOCTOU)
query flags every check→use pair as high-severity, failing the PR gate on false
positives against its threat model (concurrent adversarial access). Inline
// codeql[...] suppressions are not honoured by GitHub code scanning (verified).

Extract that suite into file-lock-reclaim-identity-guard.test.ts and paths-ignore
exactly that one file via .github/codeql/codeql-config.yml. Exclusion surface is a
single file; the rest of the lock suite and all production code — including the
lock implementation itself — keep full security-extended coverage. Add a path
here only when a fixture genuinely cannot express its scenario without tripping a
security query.
