+++ packages/core/__tests__/runbook/file-lock.test.ts
@@ -15,8 +15,6 @@ import {
   isProcessAlive,
   releaseFileLock,
   releaseFileLockSync,
-  unlinkStaleByIdentity,
-  unlinkStaleByIdentitySync,
 } from '../../src/runbook/file-lock.js';
 
 interface LockContent {
@@ -228,70 +226,11 @@ describe('file-lock', () => {
     });
   });
 
-  // The reclaim delete is guarded by dev/ino identity, compared against the
-  // STILL-OPEN handle the lock was read through. Holding that handle pins the
-  // observed inode, so a lock re-created at the same path during the
-  // read→delete window cannot reuse the inode number and slip past the guard
-  // (the concurrent-reclaimer TOCTOU). These pin that guard directly; passing
-  // the open handle is also what makes the replacement's inode deterministically
-  // distinct on inode-reusing filesystems (ext4 on CI).
-  describe('reclaim identity guard', () => {
-    it('async: removes the lock when the handle still matches the path', async () => {
-      await fs.mkdir(lockDir, { recursive: true });
-      await fs.writeFile(lockFile, 'stale', 'utf8');
-      const handle = await fs.open(lockFile, 'r');
-      try {
-        expect(await unlinkStaleByIdentity(handle, lockFile)).toBe(true);
-      } finally {
-        await handle.close();
-      }
-      await expect(fs.readFile(lockFile, 'utf8')).rejects.toMatchObject({ code: 'ENOENT' });
-    });
-
-    it('async: leaves the lock intact when it was replaced under the path', async () => {
-      await fs.mkdir(lockDir, { recursive: true });
-      await fs.writeFile(lockFile, 'stale', 'utf8');
-      // Read the stale lock through a handle and keep it open — pins the
-      // observed inode so the replacement below gets a distinct one.
-      const handle = await fs.open(lockFile, 'r');
-      try {
-        // Another process reclaims + re-acquires: same path, new (live) inode.
-        await fs.rm(lockFile);
-        await fs.writeFile(lockFile, 'live-replacement', 'utf8');
-        expect(await unlinkStaleByIdentity(handle, lockFile)).toBe(false);
-        // The live replacement must survive untouched.
-        expect(await fs.readFile(lockFile, 'utf8')).toBe('live-replacement');
-      } finally {
-        await handle.close();
-      }
-    });
-
-    it('async: treats an already-vanished lock as reclaimed', async () => {
-      await fs.mkdir(lockDir, { recursive: true });
-      await fs.writeFile(lockFile, 'stale', 'utf8');
-      const handle = await fs.open(lockFile, 'r');
-      try {
-        await fs.rm(lockFile);
-        expect(await unlinkStaleByIdentity(handle, lockFile)).toBe(true);
-      } finally {
-        await handle.close();
-      }
-    });
-
-    it('sync: leaves the lock intact when it was replaced under the path', () => {
-      fsSync.mkdirSync(lockDir, { recursive: true });
-      fsSync.writeFileSync(lockFile, 'stale', 'utf8');
-      const fd = fsSync.openSync(lockFile, 'r');
-      try {
-        fsSync.rmSync(lockFile);
-        fsSync.writeFileSync(lockFile, 'live-replacement', 'utf8');
-        expect(unlinkStaleByIdentitySync(fd, lockFile)).toBe(false);
-        expect(fsSync.readFileSync(lockFile, 'utf8')).toBe('live-replacement');
-      } finally {
-        fsSync.closeSync(fd);
-      }
-    });
-  });
+  // The reclaim identity guard suite lives in its own file
+  // (file-lock-reclaim-identity-guard.test.ts) because those tests deliberately
+  // stage open→rm→recreate races that trip CodeQL's js/file-system-race query;
+  // isolating them lets that one file be excluded from analysis while the rest of
+  // this suite stays scanned.
 
   // Sync variants share the lock-file format with the async path, so a sync
   // acquire correctly observes async-held locks (and vice versa). The tests
